refactor(runtime): extract history management to history module
Move sanitize_history, persist_error_recovery, append_tool_result,
and messages_to_chat_history into a dedicated history.rs module.
prepare_history stays on Orchestrator but delegates to the extracted
functions.

orchestrator.rs shrinks by ~175 lines. All 45 tests pass.
